cohooyo_backend/requirements.txt
alembic==1.4.3
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.4
blinker==1.4
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
-e .
colorama==0.4.4
cryptography==3.2.1
cycler==0.10.0
cymem==2.0.5
#de-core-news-lg==2.3.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-UUID==0.2
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
Mako==1.1.3
MarkupSafe==1.1.1
matplotlib==3.3.3
murmurhash==1.0.5
numpy==1.19.3
packaging==20.7
pandas==1.1.5
pg8000==1.16.6
Pillow==8.0.1
plac==1.1.3
pluggy==0.13.1
preshed==3.0.5
psycopg2-binary==2.8.6
py==1.9.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
scramp==1.2.0
seaborn==0.11.0
six==1.15.0
sklearn==0.0
spacy==2.3.4
SQLAlchemy==1.3.20
srsly==1.0.5
testing.common.database==2.0.3
testing.postgresql==1.3.0
thinc==7.4.5
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.54.1
urllib3==1.26.2
wasabi==0.8.0
Werkzeug==1.0.1
